refactor: controller의 로직을 service로 이동
