+++ app/src/main/java/nextstep/jwp/RequestMapping.java
@@ -1,4 +1,9 @@
-package nextstep.jwp.controller;
+package nextstep.jwp;
+
+import nextstep.jwp.controller.Controller;
+import nextstep.jwp.controller.DefaultController;
+import nextstep.jwp.controller.LoginController;
+import nextstep.jwp.controller.RegisterController;
 
 import java.util.HashMap;
 import java.util.Map;
