+++ app/src/main/java/nextstep/jwp/controller/RegisterController.java
@@ -1,19 +1,19 @@
 package nextstep.jwp.controller;
 
-import nextstep.jwp.db.InMemoryUserRepository;
-import nextstep.jwp.model.User;
+import nextstep.jwp.application.UserService;
 import nextstep.jwp.model.http.HttpRequest;
 import nextstep.jwp.model.http.HttpResponse;
 import nextstep.jwp.util.FileUtils;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.Optional;
 
 public class RegisterController extends AbstractController {
 
-    private static final Logger LOG = LoggerFactory.getLogger(RegisterController.class);
+    private final UserService userService;
+
+    public RegisterController() {
+        this.userService = new UserService();
+    }
 
     @Override
     protected void doGet(HttpRequest request, HttpResponse response) throws IOException {
@@ -22,13 +22,7 @@ public class RegisterController extends AbstractController {
 
     @Override
     protected void doPost(HttpRequest request, HttpResponse response) throws IOException {
-        User newUser = new User(request.getParameter("account"),
-                request.getParameter("password"),
-                request.getParameter("email"));
-        InMemoryUserRepository.save(newUser);
-
-        Optional<User> savedUser = InMemoryUserRepository.findByAccount(request.getParameter("account"));
-        LOG.debug("Saved user : {}", savedUser);
+        userService.saveUser(request);
         response.redirect("/index.html");
     }
 }
